Revert "feat/branching depositAndSwap per cellar"
This reverts commit 01be122937eeb5eecce805dafded35bfec22af7e.
