+++ src/components/_cards/PortfolioCard.tsx
@@ -218,9 +218,7 @@ export const PortfolioCard: VFC<BoxProps> = (props) => {
                       {userStakes?.totalClaimAllRewards.formatted ||
                         "..."}
                       <ImportMetamaskButton
-                        address={
-                          cellarConfig.rewardTokenAddress || ""
-                        }
+                        address={cellarConfig.rewardTokenAddress}
                       />
                     </>
                   ) : (
